Implement DSL for Cypher triple queries
While working on #52 it occurred to me that I would have to non-trivially re-write a lot of these match statements to add in type constraints. This PR adds a more generic function that is amenable to easily adding type constraints

Caveat: it doesn't have a first class notion of the `*1...` that appears in the predecessor/successor nodes. I could add that, but it might be premature generalization